Added CRUD pemohon, view senjata api, view amunisi

+++ application/controllers/Admin.php
@@ -24,30 +24,36 @@ class Admin extends MY_Controller
 
     public function index()
     {
+        $this->load->model('senjata_api_m');
+        $this->load->model('amunisi_m');
         $this->data['title']    = 'Dashboard Admin';
         $this->data['pemohon']  = $this->Pemohon_m->get();
+        $this->data['amunisi']  = $this->amunisi_m->get();
+        $this->data['senpi']    = $this->senjata_api_m->get();
         $this->data['content']  = 'admin/dashboard';
         $this->template($this->data);
     }
 
     public function pemohon()
     {
-        // $nrp = $this->POST('nrp');
-        // if(isset($nrp))
-        // {
-        //     $cek_pemohon = $this->Pemohon_m->get_row(['nrp' => $nrp]);
+        if($this->POST('change'))
+        {
+            $nrp = $this->POST('nrp');
+            $cek_pemohon = $this->Pemohon_m->get_row(['nrp' => $nrp]);
 
-        //     if ($cek_pemohon->status == '1')
-        //     {
-        //         $this->Pemohon_m->update($nrp, ['status' => '0']);
-        //         echo '<button class="btn btn-danger" onclick="changeStatus('.$nrp.')"><i class="fa fa-close"></i> Tidak</button>';
-        //     }
-        //     else
-        //     {
-        //         $this->Pemohon_m->update($nrp, ['status' => '1']);
-        //         echo '<button class="btn btn-success" onclick="changeStatus('.$nrp.')"><i class="fa fa-check"></i> Iya</button>'; 
-        //     }
-        // }
+            if ($cek_pemohon->status == '1')
+            {
+                $this->Pemohon_m->update($nrp, ['status' => '0']);
+                echo '<button class="btn btn-danger" onclick="changeStatus('.$nrp.')"><i class="fa fa-close"></i> Tidak</button>';
+                exit;
+            }
+            else
+            {
+                $this->Pemohon_m->update($nrp, ['status' => '1']);
+                echo '<button class="btn btn-success" onclick="changeStatus('.$nrp.')"><i class="fa fa-check"></i> Iya</button>'; 
+                exit;
+            }
+        }
 
         $this->load->model('senjata_api_m');
 
@@ -69,6 +75,36 @@ class Admin extends MY_Controller
             exit;
         }
 
+        if ($this->POST('edit'))
+        {
+            $this->data['pemohon'] = [
+                'nrp'           => $this->POST('nrp'),
+                'nama_anggota'  => $this->POST('nama_anggota'),
+                'pangkat'       => $this->POST('pangkat'),
+                'jabatan'       => $this->POST('jabatan'),
+                'kesatuan'      => $this->POST('kesatuan'),
+                'kelengkapan'   => $this->POST('kelengkapan'),
+                'no_senpi'      => $this->POST('no_senpi'),
+                'jumlah_amunisi'=> $this->POST('jumlah_amunisi')
+            ];
+            $this->Pemohon_m->update($this->POST('old_nrp'), $this->data['pemohon']);
+            $this->flashmsg('<i class="glyphicon glyphicon-success"></i> Data pemohon berhasil di-edit');
+            redirect('admin/pemohon');
+            exit;
+        }
+
+        if ($this->input->get('nrp'))
+        {
+            $this->data['pemohon'] = $this->Pemohon_m->get_row(['nrp' => $this->input->get('nrp')]);
+            $no_senpi = [];
+            $this->data['senpi']    = $this->senjata_api_m->get();
+            foreach ($this->data['senpi'] as $senpi)
+                $no_senpi[$senpi->no_senpi] = $senpi->no_senpi;
+            $this->data['pemohon']->dropdown = form_dropdown('no_senpi', $no_senpi, $this->data['pemohon']->no_senpi, ['class' => 'form-control']);
+            echo json_encode($this->data['pemohon']);
+            exit;
+        }
+
         if ($this->POST('delete')) 
         {
             $this->Pemohon_m->delete($this->POST('nrp'));
@@ -80,6 +116,24 @@ class Admin extends MY_Controller
         $this->data['title']    = 'Data Pemohon Senjata Api';
         $this->data['pemohon']  = $this->Pemohon_m->get();
         $this->data['content']  = 'admin/pemohon';
+        $this->template($this->data);
+    }
+
+    public function senpi()
+    {
+        $this->load->model('senjata_api_m');
+        $this->data['title']    = 'Senjata Api di Gudang';
+        $this->data['senpi']    = $this->senjata_api_m->get();
+        $this->data['content']  = 'admin/senpi';
+        $this->template($this->data);
+    }
+
+    public function stok_amunisi()
+    {
+        $this->load->model('amunisi_m');
+        $this->data['title']    = 'Stok Amunisi';
+        $this->data['amunisi']  = $this->amunisi_m->get();
+        $this->data['content']  = 'admin/amunisi';
         $this->template($this->data);
     }
 }
